refactor(web): move FileViewer JSON-number display formatter into the slice
Add the file-viewer slice's formatters.ts with the precision-safe JSON
pretty-printer, moved byte-for-byte from FileViewer.tsx:

- formatJsonFileTextForDisplay (public) plus its private helpers
  (hasPrecisionSensitiveJsonNumberText, isSignedNegativeZeroJsonNumberToken,
  isPrecisionSensitiveJsonNumberToken, parseJsonNumberTokenAsDecimal,
  normalizeDecimalParts, hasUnsafeJsonNumber, isJsonFile).
- FileViewer.tsx imports it from the slice barrel; only the public entry point
  crosses the boundary.
- tests/features/file-viewer/formatters.test.ts: 13 cases (formatters.ts
  coverage 97.3% stmts / 96.2% branch / 100% funcs / 100% lines).

Gates: pnpm guard (incl. web slice boundary check) green; FileViewer.test 180
green; formatters typecheck clean in isolation.
