+++ apps/web/src/features/file-viewer/formatters.ts
@@ -1,0 +1,115 @@
+// Pure display formatters for the file-viewer slice. No React, transport, or
+// DOM — they map raw file data to display strings and test with zero doubles.
+import type { ProjectFile } from '../../types';
+
+// Pretty-print a `.json` file for the read-only source view, but ONLY when the
+// round-trip is lossless. JSON re-serialization silently rewrites numbers, so
+// we bail back to the raw text whenever a number would change: precision-
+// sensitive literals (checked on the token text before parsing) and unsafe
+// runtime numbers (checked on the parsed value). Non-JSON files and parse
+// failures also pass through untouched.
+export function formatJsonFileTextForDisplay(file: ProjectFile, text: string): string {
+  if (!isJsonFile(file)) return text;
+  try {
+    if (hasPrecisionSensitiveJsonNumberText(text)) return text;
+    const parsed = JSON.parse(text) as unknown;
+    if (hasUnsafeJsonNumber(parsed)) return text;
+    return JSON.stringify(parsed, null, 2);
+  } catch {
+    return text;
+  }
+}
+
+// Scan the raw JSON text (outside string literals) for any number token whose
+// value would not survive a JSON.parse -> JSON.stringify round-trip: a signed
+// negative zero (`-0`), or a literal that Number() cannot represent exactly.
+function hasPrecisionSensitiveJsonNumberText(text: string): boolean {
+  let inString = false;
+  let escaped = false;
+  const numberTokenPattern = /-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?/y;
+  for (let i = 0; i < text.length;) {
+    const char = text[i];
+    if (inString) {
+      if (escaped) {
+        escaped = false;
+      } else if (char === '\\') {
+        escaped = true;
+      } else if (char === '"') {
+        inString = false;
+      }
+      i += 1;
+      continue;
+    }
+
+    if (char === '"') {
+      inString = true;
+      i += 1;
+      continue;
+    }
+
+    numberTokenPattern.lastIndex = i;
+    const match = numberTokenPattern.exec(text);
+    if (!match) {
+      i += 1;
+      continue;
+    }
+
+    const token = match[0];
+    if (isSignedNegativeZeroJsonNumberToken(token)) return true;
+    if (/[.eE]/.test(token) && isPrecisionSensitiveJsonNumberToken(token)) return true;
+    i = numberTokenPattern.lastIndex;
+  }
+  return false;
+}
+
+function isSignedNegativeZeroJsonNumberToken(token: string): boolean {
+  return /^-0(?:\.0+)?(?:[eE][+-]?\d+)?$/.test(token);
+}
+
+function isPrecisionSensitiveJsonNumberToken(token: string): boolean {
+  const parsed = Number(token);
+  if (!Number.isFinite(parsed)) return true;
+  const rendered = JSON.stringify(parsed);
+  if (!rendered) return true;
+  const originalValue = parseJsonNumberTokenAsDecimal(token);
+  const renderedValue = parseJsonNumberTokenAsDecimal(rendered);
+  return (
+    !originalValue ||
+    !renderedValue ||
+    originalValue.coefficient !== renderedValue.coefficient ||
+    originalValue.exponent !== renderedValue.exponent
+  );
+}
+
+function parseJsonNumberTokenAsDecimal(token: string): { coefficient: bigint; exponent: number } | null {
+  const match = /^(-)?(\d+)(?:\.(\d+))?(?:[eE]([+-]?\d+))?$/.exec(token);
+  if (!match) return null;
+  const [, sign, integerPart, fractionPart = '', exponentPart = '0'] = match;
+  const coefficient = BigInt(`${sign ?? ''}${integerPart}${fractionPart}`);
+  const exponent = Number(exponentPart) - fractionPart.length;
+  return normalizeDecimalParts(coefficient, exponent);
+}
+
+function normalizeDecimalParts(coefficient: bigint, exponent: number): { coefficient: bigint; exponent: number } {
+  if (coefficient === 0n) return { coefficient: 0n, exponent: 0 };
+  let normalizedCoefficient = coefficient;
+  let normalizedExponent = exponent;
+  while (normalizedCoefficient % 10n === 0n) {
+    normalizedCoefficient /= 10n;
+    normalizedExponent += 1;
+  }
+  return { coefficient: normalizedCoefficient, exponent: normalizedExponent };
+}
+
+function hasUnsafeJsonNumber(value: unknown): boolean {
+  if (typeof value === 'number') {
+    return !Number.isFinite(value) || (Number.isInteger(value) && !Number.isSafeInteger(value));
+  }
+  if (Array.isArray(value)) return value.some(hasUnsafeJsonNumber);
+  if (value && typeof value === 'object') return Object.values(value).some(hasUnsafeJsonNumber);
+  return false;
+}
+
+function isJsonFile(file: ProjectFile): boolean {
+  return file.name.toLowerCase().endsWith('.json') || file.mime.toLowerCase().startsWith('application/json');
+}
